+++ example/demo/forms.py
@@ -1,9 +1,0 @@
-from django import forms
-
-from demo.models import DemoImport
-
-
-class DemoImportForm(forms.ModelForm):
-    class Meta:
-        model = DemoImport
-        fields = ["label"]